Code refactoring and cleaning up
