+++ testdata/scenarios/custom-chart/Chart.yaml.tmpl
@@ -29,7 +29,7 @@ maintainers:
 {{if gt (len .Websites) 0 }}
 sources:
 {{- range .Websites}}
-    - {{.Url}}
+    - {{.URL}}
 {{end -}}
 {{end}}
 {{end}}
